smarthouse/requirements.txt
Django==5.0.7
psycopg2-binary==2.9.10
paho-mqtt==2.1.0